[Refactor]: java.util.date -> sql.date로 변경
